+++ src/packages/core/property-editor/index.ts
@@ -2,4 +2,4 @@ export * from './components/index.js';
 export * from './config/index.js';
 export * from './constants.js';
 export * from './events/index.js';
-export * from './property-editor-ui-picker/index.js';
+export * from './ui-picker-modal/index.js';
