+++ src/packages/core/property-editor/ui-picker-modal/index.ts
