+++ src/packages/core/property-editor/ui-picker-modal/manifests.ts
