+++ src/packages/core/property-editor/ui-picker-modal/property-editor-ui-picker-modal.stories.ts
@@ -4,7 +4,7 @@ import type { Meta, Story } from '@storybook/web-components';
 import { html } from '@umbraco-cms/backoffice/external/lit';
 
 import './property-editor-ui-picker-modal.element.js';
-import '../../../core/components/body-layout/body-layout.element.js';
+import '../../components/body-layout/body-layout.element.js';
 
 export default {
 	title: 'API/Modals/Layouts/Property Editor UI Picker',
